+++ packages/backend/src/application/usecases/auth.ts
@@ -2,8 +2,8 @@ import type { Context } from "hono";
 import { inject, injectable } from "inversify";
 import { DiscordID, User } from "../../domain/models/User";
 import { UserAuth } from "../../domain/models/UserAuth";
-import type { UserRepositoryInterface } from "../../domain/repositories/user";
-import type { UserAuthRepositoryInterface } from "../../domain/repositories/userAuth";
+import type { UserAuthRepositoryInterface } from "../../domain/repositories/UserAuthRepository";
+import type { UserRepositoryInterface } from "../../domain/repositories/UserRepository";
 import type { DiscordAuthServiceInterface } from "../../domain/services/discord-auth";
 import type { JwtServiceInterface } from "../../domain/services/jwt";
 import { TYPES } from "../../infrastructure/config/types";
